Updated up to add method.

+++ src/app/06-services/todos.component.ts
@@ -1,25 +1,25 @@
 import { TodoService } from './todo.service';
 
 export class TodosComponent {
-  todos: any[] = [];
+  todos: any = [];
   message: string = '';
 
   constructor(private service: TodoService) {}
 
   ngOnInit() {
-    //this.service.getTodos().subscribe(t => this.todos = t);
+    this.service.getTodos().subscribe((t) => (this.todos = t));
   }
 
   add() {
-    /* var newTodo = { title: '... ' };
+    var newTodo = { title: '... ' };
     this.service.add(newTodo).subscribe(
-      t => this.todos.push(t),
-      err => this.message = err); */
+      (t) => this.todos.push(t),
+      (err) => (this.message = err)
+    );
   }
 
   delete(id: number) {
-    console.log(id + this.message);
-    /* if (confirm('Are you sure?'))
-      this.service.delete(id).subscribe(); */
+    //console.log(id + this.message);
+    if (confirm('Are you sure?')) this.service.delete(id).subscribe();
   }
 }
